requirements.txt
Flask == 1.1.1
Flask_Migrate == 2.5.3
Flask_SQLAlchemy == 2.4.0
SQLAlchemy == 1.3.17
alembic == 1.4.3
flask_marshmallow == 0.14.0
marshmallow==3.8.0
marshmallow-sqlalchemy==0.23.1
pytest == 6.1.1
python-dotenv==0.14.0
psycopg2==2.8.5
requests == 2.23.0
short_url == 1.2.2
